get the code to compile and run.
Implement default constructors and destructors of the base
strategy class. Note that you cannot derive from a class with an
undeclared virtual function such as objectiveFunctionStrategy's destructor.

Also it is confusing (also for the compiler I think) to include cpp files
so i moved the ObjectiveFunctionhyperBowl definition into it's header
and included that.

tested compiling with gcc 4.7 with --std=c++11. Compiling works and code
runs without errors.